requirements.txt
feedparser
newspaper3k
requests
beautifulsoup4
lxml
